Neet to set PKG_CONFIG_PATH as env variable.
Setting it as a meson option works for meson, but g-ir-scanner also uses
pkg-config during the build and it will fail to find .pc files if the
environment variable isn't set. Doing so takes care of both use cases,
so that's preferable. It doesn't seem like the failure to find the .pc
files with g-ir-scanner breaks the build, but it's best to clear up the
error anyway.